functions/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
groq>=0.4.0
requests>=2.28.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
